dev-requirements.txt
asgiref==3.4.1
Django==4.0
django-cors-headers==3.11.0
django-filter==21.1
django-paranoid-model==1.0.7
djangorestframework==3.13.1
pytz==2021.3
PyYAML==6.0
sqlparse==0.4.2
tzdata==2021.5
uritemplate==4.1.1
psycopg2==2.9.3
